Move auto unseal test out of the enterprise test file
